+++ app/components/championship-select.tsx
@@ -1,26 +1,21 @@
 import { type ChangeEvent } from 'react';
-import { useLocation, useMatches, useNavigate } from '@remix-run/react';
-
-import { object, safeParse, string } from 'valibot';
+import { useLocation, useNavigate } from '@remix-run/react';
 
 import { useChampionships } from '#app/utils/hooks/use-championships';
+import { useViewSegment } from '#app/utils/route-handle';
 
 export function ChampionshipSelect() {
   const { championships, current } = useChampionships();
   const { search } = useLocation();
-  const matches = useMatches();
+  const viewSegment = useViewSegment();
   const navigate = useNavigate();
-
-  const viewMatch = matches.at(-1);
-  const handle = safeParse(object({ viewPath: string() }), viewMatch?.handle);
-  const viewPath = handle.success ? handle.output.viewPath : '';
 
   function handleSelect(ev: ChangeEvent<HTMLSelectElement>) {
     const slug = ev.target.value;
     const championship = slug === championships[0]?.slug ? '' : slug;
     navigate(
       {
-        pathname: `/${[championship, viewPath].filter(Boolean).join('/')}`,
+        pathname: `/${[championship, viewSegment].filter(Boolean).join('/')}`,
         search,
       },
       { unstable_viewTransition: true },