+++ app/modules/auth/auth.server.ts
@@ -1,13 +1,57 @@
-import { Authenticator } from 'remix-auth';
+import { redirect } from '@remix-run/node';
 
-import type { User } from '#db/model';
+import { invariant } from '#app/utils/invariant';
 
-import { authSession } from './auth-session.server';
-import { totpStrategy } from './auth-strategy.server';
+import { getUserByEmail, getUserById } from '../db/model/users';
+import {
+  commitSession,
+  destroySession,
+  getSession,
+} from './auth-session.server';
 
-const authenticator = new Authenticator<User>(authSession, {
-  throwOnError: true,
-});
-authenticator.use(totpStrategy, 'TOTP');
+export async function getUserId(request: Request) {
+  const session = await getSession(request.headers.get('Cookie'));
+  const userId = session.get('user:id');
+  return userId ?? null;
+}
 
-export { authenticator };
+export async function getUser(request: Request) {
+  const userId = await getUserId(request);
+
+  return userId && (await getUserById(userId));
+}
+
+export async function requireAnonymous(request: Request) {
+  const userId = await getUserId(request);
+  if (userId) {
+    throw redirect('/');
+  }
+}
+
+export async function isKnownEmail(email: string) {
+  const user = await getUserByEmail(email);
+  return user !== null;
+}
+
+export async function login(request: Request, email: string) {
+  const session = await getSession(request.headers.get('Cookie'));
+
+  const user = await getUserByEmail(email);
+  invariant(user, 'Unknown authenticated user.');
+
+  session.set('user:id', user.id);
+
+  return redirect('/', {
+    headers: {
+      'Set-Cookie': await commitSession(session),
+    },
+  });
+}
+
+export async function logout(request: Request) {
+  const session = await getSession(request.headers.get('Cookie'));
+
+  return redirect('/', {
+    headers: { 'Set-Cookie': await destroySession(session) },
+  });
+}