+++ app/root.tsx
@@ -1,9 +1,21 @@
-import { type LinksFunction } from '@remix-run/node';
+import {
+  json,
+  type LinksFunction,
+  type LoaderFunctionArgs,
+} from '@remix-run/node';
 import { Links, LiveReload, Outlet, Scripts } from '@remix-run/react';
 
 import styles from '#app/styles.css';
 
+import { getUser } from './modules/auth/auth.server';
+
 export const links: LinksFunction = () => [{ rel: 'stylesheet', href: styles }];
+
+export async function loader({ request }: LoaderFunctionArgs) {
+  const user = await getUser(request);
+
+  return json({ user });
+}
 
 export default function App() {
   return (